+++ src/lib.rs
@@ -1,15 +1,25 @@
 #![forbid(unsafe_code)]
 
-pub mod error;
 pub mod header;
 pub mod tables;
 
 use std::str;
-
-use error::Mp3Error;
 use header::*;
 
 pub static ID3V1_LEN: usize = 128;
+
+#[derive(Debug)]
+pub enum Mp3Error {
+    ID3Error, // Unable to trim ID3 tag
+    HeaderError, // Incorrect Header
+    Utf8Error(str::Utf8Error),
+}
+
+impl From<str::Utf8Error> for Mp3Error {
+    fn from(e: str::Utf8Error) -> Mp3Error {
+        Mp3Error::Utf8Error(e)
+    }
+}
 
 // Trim ID3 tag from data and find first frame
 pub fn trim_data(data: &[u8]) -> Result<&[u8], Mp3Error> {
@@ -39,9 +49,6 @@ pub fn trim_data(data: &[u8]) -> Result<&[u8], Mp3Error> {
 }
 
 pub fn frame_size(header: &FrameHeader) -> usize {
-    (144000usize * (header.bitrate() as usize)) / (header.sampling_rate() as usize)
-    + match header.padding() {
-        Padding::Yes => 1usize,
-        Padding::No  => 0usize,
-    }
+    (144000usize * (header.bitrate as usize)) / (header.sampling_rate as usize)
+    + if header.padding { 1usize } else { 0usize }
 }
